add auto accept friends

+++ bot.py
@@ -3,14 +3,7 @@ from agent import *
 import time
 import re
 from random import randint
-
-# agent = PairMatchAgent()
-# turing = Tuling(api_key='25da5f49d4ad44d48a477543f0b3f55e')
-# helper = ensure_one(bot.friends().search(helper_name))
-# # list of msgs that are not answered yet
-# msg_list = []
-# # cache for multiple line answer from helper
-# msg_cache = []
+import warnings
 
 
 def preprocess_raw_text(raw_text):
@@ -19,14 +12,14 @@ def preprocess_raw_text(raw_text):
     if raw_text.lower().startswith('godmode'):
         mode = 'god'
         text = raw_text.strip().split('godmode')[-1]
-    elif raw_text.startswith('小哥哥') or len(raw_text) < 4:
-        mode = 'chitchat'
-        text = raw_text
+    elif raw_text.startswith('请问'):
+        mode = 'qa'
+        text = raw_text.strip().split('请问')[-1]
     elif raw_text.lower().startswith('debug'):
         mode = 'debug'
         text = raw_text.strip().split('debug')[-1]
     else:
-        mode = 'qa'
+        mode = 'chitchat'
         text = raw_text.strip()
     while text[0] in skipwords and len(text) > 0:
         text = text[1:]
@@ -41,7 +34,7 @@ class WXBot(Bot):
         # agent for chitchat
         self.turing = Tuling(api_key='25da5f49d4ad44d48a477543f0b3f55e')
         # human assistant
-        self.helper = self.friends().search(helper_name)[0]
+        self.helper = ensure_one(self.friends().search(helper_name))
         # human master, won't be locked
         # self.master = ensure_one(self.friends().search(master_name))
         # list of msgs that are not answered yet
@@ -55,14 +48,18 @@ class WXBot(Bot):
         # asking
         self.asking = False
         # backend users
-        god = self.friends().search(god_name)[0]
+        god = ensure_one(self.friends().search(god_name))
         self.gods = {god}
         # waiting record decision
         self.waiting = False
         # uncommented user message dict
         self.user_msg = {}
         # bonus reminded
-        self.bonused = set()
+        self.reminded = set()
+        # statistics
+        self.stats = {'success': 0,
+                      'fail': 0,
+                      'total_msg': 0}
 
     def get_save_path(self):
         date = time.strftime("%D:%H:%M:%S ", time.localtime(time.time()))
@@ -128,10 +125,12 @@ class WXBot(Bot):
                         if content_a.startswith('PICTURE:'):
                             path = content_a.split('PICTURE:')[-1]
                             # print('os.remove(path=path)')
-                            try:
+                            if os.path.exists(path + ' '):
                                 os.remove(path=path + ' ')
-                            except:
+                            elif os.path.exists(path):
                                 os.remove(path=path)
+                            else:
+                                warnings.warn('path %s not exist' % path)
             if len(ans) > 0 and content_q:
                 keeps.append((content_q, ans))
         count = 0
@@ -160,6 +159,10 @@ class WXBot(Bot):
                 msg.chat.send('record is removed. Current record count: %s' % count)
             else:
                 msg.chat.send('record is not found. Current record count: %s' % count)
+        elif text.startswith('show_stats'):
+            msg.chat.send('success: %s' % self.stats['success'])
+            msg.chat.send('fail: %s' % self.stats['fail'])
+            msg.chat.send('total messages: %s' % self.stats['total_msg'])
 
     def reply_record(self, msg):
         if msg.text.lower() == 'y' and len(self.msg_cache) > 0:
@@ -175,8 +178,6 @@ class WXBot(Bot):
             del self.user_msg[self.msg_list[0].chat]
         if len(self.msg_cache) > 0:
             self.msg_list[0].chat.send('转自人事小姐姐的消息：')
-            self.msg_list[0].chat.send('问题： %s' % self.msg_list[0].text)
-            self.msg_list[0].chat.send('回复：')
         for msg in self.msg_cache:
             msg.forward(self.msg_list[0].chat)
         self.locked_users.remove(self.msg_list[0].chat)
@@ -225,6 +226,10 @@ class WXBot(Bot):
             self.asking = True
 
     def reply_user(self, msg):
+        # only reply text msg
+        if msg.type != 'Text':
+            return
+
         text, mode = preprocess_raw_text(msg.text.lower())
         # meaningless message
         if len(text) == 0:
@@ -238,10 +243,12 @@ class WXBot(Bot):
         # when user comment
         if msg.chat in self.user_msg and msg.text == '是':
             msg.chat.send('嘻嘻～')
+            self.stats['success'] += 1
             del self.user_msg[msg.chat]
             return
         elif msg.chat in self.user_msg and msg.text == '否':
             msg.chat.send('好吧，我再去问问人事小姐姐看')
+            self.stats['fail'] += 1
             self.forward_to_helper(self.user_msg[msg.chat])
             return
         elif msg.chat in self.user_msg:
@@ -251,9 +258,9 @@ class WXBot(Bot):
         if mode == 'chitchat':
             self.turing.do_reply(msg)
             # bonus
-            if randint(1, 8) == 1 and msg.chat not in self.bonused:
-                msg.chat.send('偷偷告诉你个秘密，消息前面带上小哥哥三个字就可以和我闲聊了。不要告诉别人哦～')
-                self.bonused.add(msg.chat)
+            if msg.chat not in self.reminded:
+                msg.chat.send('告诉你个秘密，消息前面带上请问两个字，我就会认真回答你的问题。不要告诉别人哦～')
+                self.reminded.add(msg.chat)
             return
         else:
             certainty, qa_pair = self.agent.predict(msg.text)
@@ -279,21 +286,7 @@ friends = bot.friends()
 
 @bot.register(friends)
 def reply_my_friend(msg):
-    # target = ensure_one(bot.friends().search('代号H'))
-    # if msg.chat == target and msg.type == 'Picture':
-    #     print('test forward')
-    #     msg.forward(target)
-    #     print('test record and send')
-    #     qa_pair = ('图片', [msg])
-    #     bot.add_record(qa_pair)
-    # elif msg.chat == target and msg.text == '图片':
-    #     certainty, qa_pair = bot.agent.predict(msg.text)
-    #     q, ans = qa_pair
-    #     msg.chat.send('certainty: %s' % certainty)
-    #     msg.chat.send('matched: %s' % q)
-    #     msg.chat.send('answer: ')
-    #     bot.send_user_ans(user=msg.chat, ans=ans)
-
+    bot.stats['total_msg'] += 1
     if msg.type == 'Text' and msg.text.startswith('godmode') and msg.chat in bot.gods:
         print('reply god')
         bot.reply_god(msg)
@@ -305,4 +298,13 @@ def reply_my_friend(msg):
         bot.reply_user(msg)
 
 
+@bot.register(msg_types=FRIENDS)
+def auto_accept_friends(msg):
+    # 判断好友请求中的验证文本
+    if '意能通' in msg.text.lower():
+        # 接受好友 (msg.card 为该请求的用户对象)
+        new_friend = bot.accept_friend(msg.card)
+        # 或 new_friend = msg.card.accept()
+        # 向新的好友发送消息
+        new_friend.send('你好，欢迎意能通的小伙伴～')
 embed()
